refactor: 패키지 파일 세분화
