+++ src/main/java/lotto/config/AppConfig.java
@@ -3,21 +3,21 @@ package lotto.config;
 
 import java.util.List;
 import lotto.controller.LottoGameController;
-import lotto.domain.DefaultLottoFactory;
-import lotto.domain.LottoFactory;
-import lotto.domain.LottoNumbersGenerator;
-import lotto.domain.Lottos;
+import lotto.domain.ticket.DefaultLottoFactory;
+import lotto.domain.ticket.LottoFactory;
+import lotto.domain.ticket.LottoNumbersGenerator;
+import lotto.domain.ticket.Lottos;
 import lotto.dto.LottosDto;
-import lotto.service.DtoMapper;
-import lotto.service.LottoPurchaseService;
-import lotto.service.LottoPurchaseServiceImpl;
-import lotto.service.LottoResultService;
-import lotto.service.LottoResultServiceImpl;
+import lotto.service.mapper.DtoMapper;
+import lotto.service.puchase.LottoPurchaseService;
+import lotto.service.puchase.LottoPurchaseServiceImpl;
+import lotto.service.result.LottoResultService;
+import lotto.service.result.LottoResultServiceImpl;
 import lotto.utils.parser.Parser;
 import lotto.utils.parser.StringToIntListParser;
 import lotto.utils.parser.StringToIntParser;
-import lotto.utils.validator.ComparisonValidator;
-import lotto.utils.validator.InputValidator;
+import lotto.utils.inputValidator.comparison.ComparisonValidator;
+import lotto.utils.inputValidator.InputValidator;
 import lotto.view.input.ConsoleInputView;
 import lotto.view.input.InputView;
 import lotto.view.output.ConsoleOutputView;
