+++ src/main/java/lotto/domain/result/MatchResult.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.result;
 
 public class MatchResult {
     private final int matchedNumberCount;
