+++ src/main/java/lotto/domain/ticket/LottoFactory.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.ticket;
 
 public interface LottoFactory {
     Lotto createLotto();
