+++ src/main/java/lotto/domain/ticket/LottoNumbersGenerator.java
@@ -1,16 +1,16 @@
-package lotto.domain;
+package lotto.domain.ticket;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
 
-import static lotto.constants.LottoConstants.LOWER_BOUND;
-import static lotto.constants.LottoConstants.UPPER_BOUND;
-import static lotto.constants.LottoConstants.NUMBERS_PER_LOTTO;
+import static lotto.constants.LottoTicket.LOWER_BOUND;
+import static lotto.constants.LottoTicket.UPPER_BOUND;
+import static lotto.constants.LottoTicket.NUMBERS_PER_LOTTO;
 
 
 public class LottoNumbersGenerator implements NumberGenerator {
 
-    public List<Integer> generateNumbers() {
+    public List<Integer> generate() {
         return  Randoms.pickUniqueNumbersInRange(
                 LOWER_BOUND.getValue()
                 ,UPPER_BOUND.getValue()
