+++ src/main/java/lotto/service/mapper/DtoMapper.java
@@ -1,4 +1,4 @@
-package lotto.service;
+package lotto.service.mapper;
 
 public interface DtoMapper <Entity,Dto> {
     Dto toDto(Entity entity);
