+++ src/main/java/lotto/utils/inputValidator/InputValidator.java
@@ -1,4 +1,4 @@
-package lotto.utils.validator;
+package lotto.utils.inputValidator;
 
 public interface InputValidator<Type> {
     void validate(Type value);
