+++ src/main/java/lotto/utils/inputValidator/comparison/ComparisonValidator.java
@@ -1,6 +1,7 @@
-package lotto.utils.validator;
+package lotto.utils.inputValidator.comparison;
 
 import java.util.List;
+import lotto.utils.inputValidator.InputValidator;
 
 public interface ComparisonValidator extends InputValidator<String> {
     void validateWithComparison(String validationTarget, List<Integer> comparisons);
